CameraHAL: modified device node probing
Modified the device node probing from last device node as USB Camera typically
gets registered at last.
There is one problem in probing (opening) device node starting from 0, as there
are many Video Capture drivers in kernel that are typically not camera, and the
open will succeed in them which should not be the case because they are not
Camera driver.

But, need an overall better mechanism of probing device nodes instead of opening all
device nodes in /dev/video*

Change-Id: I88f797ae12f15a11df3aae34143d343981e7dab0
Signed-off-by: Annamalai Lakshmanan <annamalai.lakshmanan@linaro.org>